requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
statsmodels==0.14.0
prophet==1.1.5

# Deep Learning
tensorflow==2.15.0
keras==2.15.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Dashboard & API
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
joblib==1.3.2

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
notebook==7.0.6

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1

# API Testing
requests==2.31.0
